requirements.txt
# Core dependencies
streamlit
pandas
numpy
requests
kafka-python
minio
flask

# Big Data & ML dependencies (sesuai arsitektur)
pyspark==3.4.0
py4j==0.10.9.7
scikit-learn
matplotlib
seaborn
plotly

# Additional for Spark
findspark